Refactored TracerCallback, moved `trace_model_from_checkpoint` function, added couple of checks for `logdir`.
