+++ catalyst/contrib/dl/callbacks/trace.py
@@ -1,11 +1,10 @@
 from typing import Union  # isort:skip
-from operator import lt, gt
 
 import warnings
 
 from pathlib import Path
 
-from catalyst.dl import Callback, CallbackOrder, State
+from catalyst.dl import Callback, CallbackOrder, CallbackNode, State
 from catalyst.dl.utils import (
     trace_model_from_state,
     save_traced_model,
@@ -16,35 +15,37 @@ class TracerCallback(Callback):
 
     def __init__(
             self,
-            metric_key: str,
-            stage: str = None,
-            mode: str = "max",
+            metric: str = "loss",
+            minimize: bool = True,
+            min_delta: float = 1e-6,
             method_name: str = "forward",
             requires_grad: bool = False,
             opt_level: str = None,
-            loader: Union[str, int] = None,
             trace_mode: str = "eval",
             out_dir: Union[str, Path] = None,
             out_model: Union[str, Path] = None,
     ):
         """
         Traces model using created experiment and runner.
-        
+
         Args:
-            metric_key (str): Metric key we should trace model based on
-            mode (str): Metric max or min value affects tracing.
+            metric (str): Metric key we should trace model based on
+            minimize (bool): Whether do we minimize metric or not
             method_name (str): Model's method name that will be
                 used as entrypoint during tracing
-            stage (str): Stage from experiment from which model and loader
-                will be taken
-            loader (str): Loader name to get the batch from
             trace_mode (str): Mode for model to trace
                 (``train`` or ``eval``)
             requires_grad (bool): Flag to use grads
             opt_level (str): AMP FP16 init level
-            out_dir (str): Directory to save model to
-            out_model: Path to save model to (override out_dir)
+            out_dir (Union[str, Path]): Directory to save model to
+            out_model (Union[str, Path]): Path to save model to
+                (overrides `out_dir` argument)
         """
+        super(TracerCallback, self).__init__(
+            order=CallbackOrder.External,
+            node=CallbackNode.All
+        )
+
         if trace_mode not in ["train", "eval"]:
             raise ValueError(
                 f"Unknown trace_mode '{trace_mode}'. "
@@ -58,53 +59,37 @@ class TracerCallback(Callback):
                 stacklevel=2,
             )
 
-        if mode == "max":
-            self.compare_fn = gt
-            self.default_value = float('-inf')
-        elif mode == "min":
-            self.compare_fn = lt
-            self.default_value = float('inf')
-        else:
-            raise ValueError(f"Unknown mode '{mode}. Must be 'max' or 'min'")
+        self.metric = metric
 
-        self.metric_key = metric_key
+        self.best_score = None
+        self.is_better = None
+        if minimize:
+            self.is_better = lambda score, best: score <= (best - min_delta)
+        else:
+            self.is_better = lambda score, best: score >= (best + min_delta)
 
         self.requires_grad = requires_grad
         self.method_name = method_name
         self.trace_mode = trace_mode
         self.opt_level = None
-        self.stage = stage
 
-        if out_model is not None and not isinstance(out_model, Path):
+        if out_model is not None:
             out_model = Path(out_model)
-
-        if out_dir is not None and not isinstance(out_model, Path):
-            out_dir = Path(out_dir)
-
         self.out_model = out_model
+
+        if out_dir is not None:
+            out_dir = Path(out_dir)
         self.out_dir = out_dir
-
-        if loader is None:
-            loader = 0
-        self.loader = loader
-
-        self.value = self.default_value
-
-        super(TracerCallback, self).__init__(CallbackOrder.External)
-
-    def on_epoch_start(self, state: State):
-        self.value = self.default_value
 
     def on_epoch_end(self, state: State):
 
-        if self.stage is not None and state.stage_name != self.stage:
-            return
+        score = state.valid_metrics[self.metric]
 
-        value = state.valid_metrics[self.metric_key]
+        if self.best_score is None:
+            self.best_score = score
 
-        if self.compare_fn(value, self.value):
-
-            self.value = value
+        if self.is_better(score, self.best_score):
+            self.best_score = score
 
             if self.opt_level is not None:
                 device = "cuda"
@@ -114,7 +99,6 @@ class TracerCallback(Callback):
             traced_model = trace_model_from_state(
                 state=state,
                 method_name=self.method_name,
-                loader=self.loader,
                 mode=self.trace_mode,
                 requires_grad=self.requires_grad,
                 opt_level=self.opt_level,
