+++ catalyst/dl/runner/core.py
@@ -325,6 +325,8 @@ class Runner(_StageBasedRunner):
             device (Device): Torch device or a string
             predict_params (dict): additional parameters for model forward
         """
+        assert logdir is not None, "`logdir` parameter should be specified"
+
         if batch is None:
             if loader is None:
                 raise ValueError(
@@ -378,11 +380,6 @@ class Runner(_StageBasedRunner):
             predict_params=predict_params,
         )
 
-        # Restore previous state of the model
-        getattr(model, "train" if _is_training else "eval")()
-        utils.set_requires_grad(model, _requires_grad)
-        model.to(_device)
-
         utils.save_traced_model(
             model=traced_model,
             logdir=logdir,
@@ -391,6 +388,11 @@ class Runner(_StageBasedRunner):
             requires_grad=requires_grad,
             opt_level=opt_level
         )
+
+        # Restore previous state of the model
+        getattr(model, "train" if _is_training else "eval")()
+        utils.set_requires_grad(model, _requires_grad)
+        model.to(_device)
 
         return traced_model
